+++ src/index.js
@@ -4,6 +4,7 @@ import ReactDOM from 'react-dom'
 import './index.scss'
 import 'bootstrap/dist/css/bootstrap.min.css'
 import '@fortawesome/fontawesome-free/css/all.css'
+import 'react-bootstrap-table-next/dist/react-bootstrap-table2.min.css'
 
 import App from './App'